+++ docs/release_notes.rst
@@ -5,6 +5,10 @@ Being an experimental research framework, Mitsuba 3 does not strictly follow the
 `Semantic Versioning <https://semver.org/>`_ convention. That said, we will
 strive to document breaking API changes in the release notes below.
 
+Incoming changes
+----------------
+
+- Allow extending ``AdjointIntegrator`` in Python
 
 Mitsuba 3.2.1
 -------------